requirements.txt
Pillow==9.3.0
selenium==3.141.0
urllib3==1.26.12
